Add paddings to icon and move padding changes to helpers class
